RSE-271: Rename upgraders to be in same. Adjust Upgrader number.
