+++ CRM/Civicase/Upgrader/Steps/Step0006.php
@@ -1,0 +1,20 @@
+<?php
+
+use CRM_Civicase_Setup_UpdateMenuLinks as UpdateMenuLinks;
+
+/**
+ * Updates the Manage Cases Menu URLs.
+ */
+class CRM_Civicase_Upgrader_Steps_Step0006 {
+
+  /**
+   * Updates the Manage Cases Menu URLs.
+   */
+  public function apply() {
+    $step = new UpdateMenuLinks();
+    $step->apply();
+
+    return TRUE;
+  }
+
+}
